refactor: Simplify project structure from FSD to basic layout
