+++ app/event/new/Step1.tsx
@@ -1,10 +1,9 @@
-import { Button } from '@/app/shared/components/Button/Button'
-import { DatePicker } from '@/app/shared/components/DatePicker/DatePicker'
-import { DaySelectBox } from '@/app/shared/components/DaySelectBox/DaySelectBox'
-import { SelectBox } from '@/app/shared/components/SelectBox/SelectBox'
-import { TextInputBox } from '@/app/shared/components/TextInputBox/TextInputBox'
-import { WideButton } from '@/app/shared/components/WideButton/WideButton'
-import { eventState } from '@/app/shared/store/eventState'
+import { Button } from '@/components/Button/Button'
+import { DatePicker } from '@/components/DatePicker/DatePicker'
+import { DaySelectBox } from '@/components/DaySelectBox/DaySelectBox'
+import { SelectBox } from '@/components/SelectBox/SelectBox'
+import { TextInputBox } from '@/components/TextInputBox/TextInputBox'
+import { eventState } from '@/store/eventState'
 import React, { useState } from 'react'
 import { useRecoilState } from 'recoil'
