+++ components/IconTextButton/IconTextButton.stories.ts
@@ -22,7 +22,7 @@ type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
   args: {
-    icon: '/create_event.svg',
+    icon: '/images/create_event.svg',
     alt: 'create event icon',
     title: '빠르게 만들기',
     label: '로그인없이 이벤트 링크 또는 코드로 일정을 생성하고 관리하세요.',
@@ -34,7 +34,7 @@ export const Default: Story = {
 
 export const JoinWithCode: Story = {
   args: {
-    icon: '/join_with_code.svg',
+    icon: '/images/join_with_code.svg',
     alt: 'join event icon',
     title: '코드로 참여하기',
     label: '로그인 없이 가지고 있는 링크 또는 코드로 이벤트에 참여하세요.',
